Continue to Refactor CNC_Functions
Move a bunch of things from CNC_Functions to system.cpp and gcode.cpp in the
process of refactoring CNC_Functions.  I did my best to avoid changing any code
but I either had to alter the following or couldn't resist fixing it now:

In order to allow defining empty global Axes before intializing them, moved
intialization to setup().  This also means altering MotorGearboxEncoder and
Encoder.

Rename ringbuffer to incSerialBuffer to be more descriptive.

Move rcvdMotorSettings and rcvdKinematicsSettings to sys struct so that they are
easier to track.

This still does not compile.  More work is still necessary.
