+++ cnc_ctrl_v1/config.h
@@ -21,3 +21,8 @@
                            // to complete before being interrupted, helpful for loop
                            // LOOPINTERVAL tuning
 #define LOOPINTERVAL 10000 // What is the frequency of the PID loop in microseconds
+
+// The EEPROM locations for each axis data
+#define LEFT_EEPROM_ADR     5
+#define RIGHT_EEPROM_ADR  105
+#define Z_EEPROM_ADR      205
