+++ cnc_ctrl_v1/system.h
@@ -21,10 +21,22 @@ Copyright 2014-2017 Bar Smith*/
 typedef struct {
   byte stop;                  // Stop flag.
   byte pause;                 // Pause flag.
+  byte rcvdMotorSettings;     // Motor Settings Status Flag
+  byte rcvdKinematicSettings; // Kinematics Settings Status Flag
   float position[3];          // Cartessian position of axes
   float steps[3];             // Encoder position of axes                     
 } system_t;
 extern system_t sys;
+extern Axis leftAxis;
+extern Axis rightAxis;
+extern Axis zAxis;
+extern RingBuffer incSerialBuffer;
 
+bool machineReady();
+void finalizeMachineSettings();
+void  calibrateChainLengths(String);
+void  updateKinematicsSettings(const String&);
+void  setupAxes();
+int   getPCBVersion();
 
 #endif
